+++ printf/ft_vprints.c
@@ -26,10 +26,10 @@ int	ft_vprints(const char *format, int i, va_list list)
 	if (format[i] == 'p')
 		return (ft_printp(va_arg(list, void *)));
 	if (format[i] == 'x')
-		return (ft_puthexn(va_arg(list, unsigned long int),
+		return (ft_puthexn(va_arg(list, unsigned long long int),
 				"0123456789abcdef"));
 	if (format[i] == 'X')
-		return (ft_puthexn(va_arg(list, unsigned long int),
+		return (ft_puthexn(va_arg(list, unsigned long long int),
 				"0123456789ABCDEF"));
 	if (format[i] == '%')
 	{
